modified:   .gitignore 	modified:   bot.py 	renamed:    check_meeting_time.py -> check_meeting_time/check_meeting_time.py 	renamed:    database.py -> database/database.py 	modified:   poetry.lock 	modified:   pyproject.toml
