+++ .gitignore
@@ -160,5 +160,5 @@ cython_debug/
 #.idea/
 
 log/
-database.db
+database/database.db
 *.service
