+++ pyproject.toml
@@ -14,6 +14,7 @@ dokuwiki = "^1.3.3"
 discord = "^2.3.2"
 aiocron = "^1.8"
 pytest = "^7.4.0"
+arrow = "^1.2.3"
 
 
 [build-system]
